zadania/Z0202/requirements.txt
mcp>=1.2.0
python-dotenv>=1.0.0
requests
mcp[cli]
Pillow
